pyproject.toml
[project]
name = "ruyi-packaging"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ['PyGithub']
